+++ src/Api/Package/RootPackageFileManager.php
@@ -58,13 +58,13 @@ interface RootPackageFileManager extends ConfigFileManager
      * Installs a plugin class.
      *
      * The plugin class must be passed as fully-qualified name of a class that
-     * implements {@link ManagerPlugin}. Plugin constructors must not have
+     * implements {@link PuliPlugin}. Plugin constructors must not have
      * mandatory parameters.
      *
      * @param string $pluginClass The fully qualified plugin class name.
      *
      * @throws InvalidConfigException If a class is not found, is not a class,
-     *                                does not implement {@link ManagerPlugin}
+     *                                does not implement {@link PuliPlugin}
      *                                or has required constructor parameters.
      */
     public function installPluginClass($pluginClass);