+++ src/Puli.php
@@ -12,8 +12,6 @@
 namespace Puli\RepositoryManager;
 
 use Psr\Log\LoggerInterface;
-use Puli\Repository\ResourceRepository;
-use Puli\Repository\ResourceRepositoryInterface;
 use Puli\RepositoryManager\Api\Config\ConfigFileManager;
 use Puli\RepositoryManager\Api\Discovery\DiscoveryManager;
 use Puli\RepositoryManager\Api\Environment\GlobalEnvironment;
@@ -25,6 +23,7 @@ use Puli\RepositoryManager\Api\Package\PackageManager;
 use Puli\RepositoryManager\Api\Package\PackageState;
 use Puli\RepositoryManager\Api\Package\RootPackageFileManager;
 use Puli\RepositoryManager\Api\Repository\RepositoryManager;
+use Puli\RepositoryManager\Assert\Assert;
 use Puli\RepositoryManager\Config\ConfigFileManagerImpl;
 use Puli\RepositoryManager\Config\ConfigFileStorage;
 use Puli\RepositoryManager\Config\ConfigJsonReader;
@@ -43,50 +42,78 @@ use Symfony\Component\EventDispatcher\EventDispatcher;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 /**
- * Creates environment and manager instances.
+ * Provides access to the Puli managers.
  *
- * Use this class to bootstrap the managers provided by this package:
+ * Use this class to access the managers provided by this package:
  *
  * ```php
- * $factory = new ManagerFactory();
- * $environment = $factory->createProjectEnvironment(getcwd());
- * $packageManager = $factory->createPackageManager($environment);
+ * $puli = new Puli(getcwd());
+ * $packageManager = $puli->getPackageManager();
  * ```
  *
- * To create one of the managers, you first need to create the environment that
- * the manager will work with. There are two kinds of environments:
+ * The `Puli` class either operates in the global or a project environment:
  *
  *  * The "global environment" is not tied to a specific root package. A global
  *    environment only loads the settings of the "config.json" file in the home
- *    directory. Global environments are created with
- *    {@link createGlobalEnvironment()}.
- *
+ *    directory. The `Puli` class operates in the global environment if no
+ *    project root directory is passed to the constructor. In the global
+ *    environment, only the global config file manager is available.
  *  * The "project environment" is tied to a specific Puli project. You need to
- *    pass the path to the project to the factory. The local environment will
- *    then give access to both global settings set in "config.json" in the
- *    home directory and local settings stored in "puli.json" of the project.
- *    Local environments are created with {@link createProjectEnvironment()}.
+ *    pass the path to the project's root directory to the constructor or to
+ *    {@link setRootDir()}. The configuration of the "puli.json" file in the
+ *    root directory is used to configure the managers.
  *
- * The factory creates four kinds of managers:
+ * The `Puli` class creates four kinds of managers:
  *
  *  * The "config file manager" allows you to modify entries of the
- *    "config.json" file in the home directory. Config file managers are
- *    created with {@link createConfigFileManager()}.
- *
+ *    "config.json" file in the home directory.
  *  * The "package file manager" manages modifications to the "puli.json" file
- *    of a Puli project. Use {@link createRootPackageFileManager()} to create it.
- *
+ *    of a Puli project.
  *  * The "package manager" manages the package repository of a Puli project.
- *    A package manager can be created with {@link createPackageManager()}.
- *
  *  * The "repository manager" manages the resource repository of a Puli
- *    project. It can be created with {@link createRepositoryManager()}.
+ *    project.
+ *  * The "discovery manager" manages the resource discovery of a Puli project.
  *
  * @since  1.0
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
-class ManagerFactory
+class Puli
 {
+    /**
+     * @var string|null
+     */
+    private $rootDir;
+
+    /**
+     * @var GlobalEnvironment|ProjectEnvironment
+     */
+    private $environment;
+
+    /**
+     * @var ConfigFileManager
+     */
+    private $configFileManager;
+
+    /**
+     * @var RootPackageFileManager
+     */
+    private $rootPackageFileManager;
+
+    /**
+     * @var PackageManager
+     */
+    private $packageManager;
+
+    /**
+     * @var RepositoryManager
+     */
+    private $repositoryManager;
+
+    /**
+     * @var DiscoveryManager
+     */
+    private $discoveryManager;
+
     /**
      * @var ConfigFileStorage|null
      */
@@ -96,6 +123,11 @@ class ManagerFactory
      * @var PackageFileStorage[]
      */
     private $packageFileStorages;
+
+    /**
+     * @var LoggerInterface
+     */
+    private $logger;
 
     /**
      * Parses the system environment for a home directory.
@@ -120,6 +152,209 @@ class ManagerFactory
     }
 
     /**
+     * Creates a new instance for the given Puli project.
+     *
+     * @param string $rootDir The root directory of the Puli project. If none
+     *                        is passed, the object operates in the global
+     *                        environment. You can set or switch root
+     *                        directories anytime by calling {@link setRootDir()}.
+     *
+     * @see Puli, setRootDir()
+     */
+    public function __construct($rootDir = null)
+    {
+        Assert::nullOrString($rootDir, 'The root directory must be a string or null. Got: %s');
+
+        $this->setRootDir($rootDir);
+    }
+
+    /**
+     * Sets the managed Puli project.
+     *
+     * If a previous project was set, all managers are discarded. You can also
+     * switch to the global environment by passing `null`.
+     *
+     * @param string|null $rootDir The root directory of the Puli project. If
+     *                             you pass `null`, the object operates in the
+     *                             global environment. You can set or switch
+     *                             root directories anytime.
+     */
+    public function setRootDir($rootDir)
+    {
+        Assert::nullOrString($rootDir, 'The root directory must be a string or null. Got: %s');
+
+        if ($rootDir !== $this->rootDir) {
+            $this->rootDir = $rootDir;
+
+            $this->resetEnvironment();
+        }
+    }
+
+    /**
+     * Returns the root directory of the managed Puli project.
+     *
+     * If no Puli project is managed at the moment, `null` is returned.
+     *
+     * @return string|null The root directory of the managed Puli project or
+     *                     `null` if none is set.
+     */
+    public function getRootDir()
+    {
+        return $this->rootDir;
+    }
+
+    public function setLogger(LoggerInterface $logger)
+    {
+        $this->logger = $logger;
+    }
+
+    public function getLogger()
+    {
+        return $this->logger;
+    }
+
+    /**
+     * Returns the environment.
+     *
+     * The home directory is read from the environment variable "PULI_HOME".
+     * If this variable is not set, the home directory defaults to:
+     *
+     *  * `$HOME/.puli` on Linux, where `$HOME` is the environment variable
+     *    "HOME".
+     *  * `$APPDATA/Puli` on Windows, where `$APPDATA` is the environment
+     *    variable "APPDATA".
+     *
+     * If none of these variables can be found, an exception is thrown.
+     *
+     * A .htaccess file is put into the home directory to protect it from web
+     * access.
+     *
+     * @return GlobalEnvironment|ProjectEnvironment The environment.
+     */
+    public function getEnvironment()
+    {
+        if (!$this->environment) {
+            $this->initEnvironment();
+        }
+
+        return $this->environment;
+    }
+
+    /**
+     * Returns the configuration file manager.
+     *
+     * @return ConfigFileManager The created configuration file manager.
+     */
+    public function getConfigFileManager()
+    {
+        if (!$this->environment) {
+            $this->initEnvironment();
+        }
+
+        return $this->configFileManager;
+    }
+
+    /**
+     * Returns the root package file manager.
+     *
+     * @return RootPackageFileManager The created package file manager.
+     */
+    public function getRootPackageFileManager()
+    {
+        if (!$this->environment) {
+            $this->initEnvironment();
+        }
+
+        return $this->rootPackageFileManager;
+    }
+
+    /**
+     * Returns the package manager.
+     *
+     * @return PackageManager The package manager.
+     */
+    public function getPackageManager()
+    {
+        if (!$this->environment) {
+            $this->initEnvironment();
+        }
+
+        return $this->packageManager;
+    }
+
+    /**
+     * Returns the resource repository manager.
+     *
+     * @return RepositoryManager The repository manager.
+     */
+    public function getRepositoryManager()
+    {
+        if (!$this->environment) {
+            $this->initEnvironment();
+        }
+
+        return $this->repositoryManager;
+    }
+
+    /**
+     * Returns the resource discovery manager.
+     *
+     * @return DiscoveryManager The discovery manager.
+     */
+    public function getDiscoveryManager()
+    {
+        if (!$this->environment) {
+            $this->initEnvironment();
+        }
+
+        return $this->discoveryManager;
+    }
+
+    private function resetEnvironment()
+    {
+        $this->environment = null;
+        $this->configFileManager = null;
+        $this->rootPackageFileManager = null;
+        $this->packageManager = null;
+        $this->repositoryManager = null;
+        $this->discoveryManager = null;
+    }
+
+    private function initEnvironment()
+    {
+        if ($this->rootDir) {
+            $this->initProjectEnvironment($this->rootDir);
+        } else {
+            $this->initGlobalEnvironment();
+        }
+    }
+
+    private function initGlobalEnvironment()
+    {
+        $this->environment = $this->createGlobalEnvironment();
+        $this->configFileManager = $this->environment->getHomeDirectory()
+            ? $this->createConfigFileManager($this->environment)
+            : null;
+        $this->rootPackageFileManager = null;
+        $this->packageManager = null;
+        $this->repositoryManager = null;
+        $this->discoveryManager = null;
+    }
+
+    private function initProjectEnvironment($rootDir)
+    {
+        // Create all managers and bind them to the event dispatcher
+        $this->environment = $this->createProjectEnvironment($rootDir);
+        $this->configFileManager = $this->environment->getHomeDirectory()
+            ? $this->createConfigFileManager($this->environment)
+            : null;
+        $this->rootPackageFileManager = $this->createRootPackageFileManager($this->environment);
+        $this->packageManager = $this->createPackageManager($this->environment);
+        $this->repositoryManager = $this->createRepositoryManager($this->environment, $this->packageManager);
+        $this->discoveryManager = $this->createDiscoveryManager($this->environment, $this->packageManager, $this->logger);
+    }
+
+    /**
      * Creates the global environment.
      *
      * The home directory is read from the environment variable "PULI_HOME".
@@ -137,7 +372,7 @@ class ManagerFactory
      *
      * @return GlobalEnvironment The global environment.
      */
-    public function createGlobalEnvironment()
+    private function createGlobalEnvironment()
     {
         $dispatcher = new EventDispatcher();
 
@@ -173,7 +408,7 @@ class ManagerFactory
      * @throws FileNotFoundException If the path does not exist.
      * @throws NoDirectoryException If the path points to a file.
      */
-    public function createProjectEnvironment($rootDir)
+    private function createProjectEnvironment($rootDir)
     {
         $dispatcher = new EventDispatcher();
 
@@ -195,7 +430,7 @@ class ManagerFactory
      *
      * @return ConfigFileManager The created configuration file manager.
      */
-    public function createConfigFileManager(GlobalEnvironment $environment)
+    private function createConfigFileManager(GlobalEnvironment $environment)
     {
         return new ConfigFileManagerImpl(
             $environment,
@@ -210,7 +445,7 @@ class ManagerFactory
      *
      * @return RootPackageFileManager The created package file manager.
      */
-    public function createRootPackageFileManager(ProjectEnvironment $environment)
+    private function createRootPackageFileManager(ProjectEnvironment $environment)
     {
         return new RootPackageFileManagerImpl(
             $environment,
@@ -226,7 +461,7 @@ class ManagerFactory
      *
      * @return PackageManager The package manager.
      */
-    public function createPackageManager(ProjectEnvironment $environment)
+    private function createPackageManager(ProjectEnvironment $environment)
     {
         return new PackageManagerImpl(
             $environment,
@@ -238,14 +473,12 @@ class ManagerFactory
      * Creates a resource repository manager.
      *
      * @param ProjectEnvironment $environment    The project environment.
-     * @param PackageManager     $packageManager The package manager. Optional.
+     * @param PackageManager     $packageManager The package manager.
      *
      * @return RepositoryManager The repository manager.
      */
-    public function createRepositoryManager(ProjectEnvironment $environment, PackageManager $packageManager = null)
+    private function createRepositoryManager(ProjectEnvironment $environment, PackageManager $packageManager)
     {
-        $packageManager = $packageManager ?: $this->createPackageManager($environment);
-
         return new RepositoryManagerImpl(
             $environment,
             $packageManager->getPackages(PackageState::ENABLED),
@@ -257,15 +490,13 @@ class ManagerFactory
      * Creates a resource discovery manager.
      *
      * @param ProjectEnvironment $environment    The project environment.
-     * @param PackageManager     $packageManager The package manager. Optional.
-     * @param LoggerInterface    $logger         The logger. Optional.
+     * @param PackageManager     $packageManager The package manager.
+     * @param LoggerInterface    $logger         The logger.
      *
      * @return DiscoveryManager The discovery manager.
      */
-    public function createDiscoveryManager(ProjectEnvironment $environment, PackageManager $packageManager = null, LoggerInterface $logger = null)
+    private function createDiscoveryManager(ProjectEnvironment $environment, PackageManager $packageManager, LoggerInterface $logger = null)
     {
-        $packageManager = $packageManager ?: $this->createPackageManager($environment);
-
         return new DiscoveryManagerImpl(
             $environment,
             $packageManager->getPackages(PackageState::ENABLED),